fix(threads): use WFE+SEV on multicore instead of WFI
If an operation on one core causes a thread to be added to the runqueue,
it has to be signaled to the other core in case that the scheduler on
that core is currently waiting (until now with an WFI).
This commit changes the WFI in the scheduler to WFE, so that an SEV can
be used to wake up the scheduler from the other core.