refactor(threads): add to rq and trigger scheduler in `set_state`
Use `set_state` as central function that is responsible for re-adding
threads to the runqueue if they change back to `ThreadState::Running`,
and trigger the scheduler.

This reverts e8f8410 and parts of 964c2bc. 964c2bc assumed that
`schedule` can not be called from inside a critical section because then
the whole scheduler would run inside the cs. This assumption was wrong,
because the cs implementation also disables interrupts, thus the
scheduler would anyway only be triggered after leaving the cs.

e8f8410 was done to slighly improve the performance because outside of
`set_state` we already know if the scheduler should be triggered or not
and thus an extra check could be avoided. But this is error-prone
because cases can easily be missed. So having one central method seems
like the better solution for now.