fix(threads): don't call `schedule` on thread creation
Calling `schedule` after first creating a thread is problematic during
startup, because the threads are created before threading is properly
initialized in `start_threading`.
For now, fix this by simply not calling `schedule` when a thread is
created from an Invalid state. Long term a different fix will be needed,
because threads could also be created dynamically during runtime.