feat(threads): separate critical_section implementation for kernel
Add an own implementation of a critical section, that is used on
multicore RP2040 for all critical sections in the kernel.
This prevents that independent critical sections in the user application
block the scheduler.

The implementation for the RP2040 is inspired by the `embassy_rp`
implementation of a critical section, but using a different hardware
spinlock and doesn't support nested locking.
The latter is okay because the cs is only internal and we never do
nested locking.