feat(thread): protect spinlock access as whole with cs
Before, locking and unlocking of a `Spinlock` was protected by a
critical section. Operations on the acquried `SpinlockGuard{Mut}` were
outside of a cs.
Because the overhead, and because access to `Spinlocks` are typically
very short, we now simply wrap the access to the inner type in a global
critical section.