feat(threads): create idle threads for multicore
Create one idle thread for each core when the `multicore` feature is
enabled. Thus, there will never be the case on multicore where the
runqueue is empty and we have to `wfi` in the scheduler loop.

This allows us to revert 5ab2104 and dedd887 (see commit messages of
those commits).