chore(rustc-lint): set `unsafe_op_in_unsafe_fn` to `deny`
`unsafe_op_in_unsafe_fn` is expected to be set to `warn` in the 2024
Rust edition [1], thus asserting that the previously accepted usage of
the `unsafe` keyword on functions allowing the use of `unsafe`
operations without `unsafe` blocks is deprecated.

We therefore enforce this by setting this lint to `deny` instead of
`warn`.

[1]: https://doc.rust-lang.org/edition-guide/rust-2024/unsafe-op-in-unsafe-fn.html